Transfer functions and constants to the files
Перенес функции и константы в отдельные файлы
Добавил константы на тексты и клавиатуры
Добавил функцию обработки ввода - выводит новое подменю с названием игры и кнопку возврата
Добавил команду help
